requirements.txt
py-cord==2.0.0b1

requests~=2.26.0
Pillow~=9.2.0